// Engine/Source/Programs/Horde/HordeDashboard/src/backend/Api.ts
export type ProjectData = GetProjectResponse & {
	streams?: StreamData[];
}

export type StreamData = GetStreamResponse & {
	project?: ProjectData;
	// full path as returned by server
	fullname?: string;

}
// ...
export type GetProjectCategoryResponse = {
	/**Heading for this column */
	name: string;

	/**Index of the row to display this category on */
	row: number;

	/**Whether to show this category on the nav menu */
	showOnNavMenu: boolean;

	/**Patterns for stream names to include */
	includePatterns: string[];

	/**Patterns for stream names to exclude */
	excludePatterns: string[];

	/**Streams to include in this category */
	streams: string[];
}

/**Response describing a project */
export type GetProjectResponse = {
	/**Unique id of the project */
	id: string;

	/**Name of the project */
	name: string;

	/**Order to display this project on the dashboard */
	order: number;

	/**List of streams that are in this project */
	// streams?: ProjectStreamData[];

	/**List of stream categories to display */
	categories?: GetProjectCategoryResponse[];

	/**Properties for this project. */
	properties: { [key: string]: string };

	/**Custom permissions for this object */
	acl?: GetAclResponse;
}
// ...
export type GetAgentTypeResponse = {

	/**Pool of agents to use for this agent type */
	pool: string;

	/**Name of the workspace to sync */
	workspace?: string;

	/**Pool of agents to use to execute this work */
	requirements?: GetAgentRequirementsResponse;

	/**Path to the temporary storage dir */
	tempStorageDir?: string;

	/**Environment variables to be set when executing the job */
	environment?: { [key: string]: string };

}

export type GetAgentTelemetrySampleResponse = {
	time: Date;
	userCpu: number;
	idleCpu: number;
	systemCpu: number;
	freeRam: number;
	usedRam: number;
	totalRam: number;
	workingFreeDisk: number;
	workingTotalDisk: number;
	systemFreeDisk: number;
	systemTotalDisk: number;
}

/**Information about a workspace type */
export type GetWorkspaceTypeResponse = {

	/**The Perforce server and port (eg. perforce:1666) */
	serverAndPort?: string;

	/**User to log into Perforce with (defaults to buildmachine) */
	userName?: string;

	/**Identifier to distinguish this workspace from other workspaces. Defaults to the workspace type name. */
	identifier?: string;

	/**Override for the stream to sync */
	stream?: string;

	/**Custom view for the workspace */
	view?: string[];

	/**Whether to use an incrementally synced workspace */
	incremental: boolean;
}
// ...
export type GetTemplateRefResponse = GetTemplateResponseBase & {

	/**Unique id of this template ref, (sanitized name) */
	id: string;

	/**Hash of the template definition */
	hash: string;

	/// The schedule for this ref
	schedule?: GetScheduleResponse;

	/** Step state for template in stream */
	stepStates?: GetTemplateStepStateResponse[];

	/** List of queries for the default changelist */
	defaultChange?: ChangeQueryConfig[];

	/** Whether the current user can run the template or not */
	canRun?: boolean;

}

/** Specifies defaults for running a preflight */
export type DefaultPreflightConfig = {

	/** The template id to query */
	templateId?: string;

	change?: ChangeQueryConfig;
}


/**  Configuration for an issue workflow */
export type WorkflowConfig = {

	/** Identifier for this workflow */
	id: string;

	/** Name of the tab to post summary data to */
	summaryTab?: string;

	/** Channel where reports are sent */
	reportChannel?: string;

	/** Channel where items to triage are sent */
	triageChannel?: string;
}

/**Response describing a stream alias */
export type GetStreamAliasResponse = {

	/**The unique name/identifier for this alias */
	name: string;

	/**The subset of template IDs available through this alias */
	templates: string[];

}

/**Response describing a stream */
export type GetStreamResponse = {

	/**Unique id of the stream */
	id: string;

	/**Unique id of the project containing this stream */
	projectId: string;

	/**Name of the stream */
	name: string;

	/**The config file path on the server*/
	configPath?: string;

	/**Revision of the config file */
	configRevision?: string;

	/**List of tabs to display for this stream*/
	tabs: GetStreamTabResponse[];

	/** Default template for running preflights */
	defaultPreflight?: DefaultPreflightConfig;

	/**Map of agent name to type */
	agentTypes: { [key: string]: GetAgentTypeResponse };

	/**Map of workspace name to type */
	workspaceTypes?: { [key: string]: GetWorkspaceTypeResponse };

	/**Templates for jobs in this stream */
	templates: GetTemplateRefResponse[];

	/**Properties for this stream */
	properties: { [key: string]: string };

	/** Workflows for this stream */
	workflows: WorkflowConfig[];

	/** Stream aliases for this stream */
	aliases?: GetStreamAliasResponse[];

	/** Custom permissions for this object */
	acl?: GetAclResponse;

}
// ...
export type GetAclResponse = {
	/**Entries to replace the existing ACL */
	entries: GetAclEntryResponse[];

	/**Whether to inherit permissions from the parent entity */
	inheritPermissions: boolean;
}
// ...
export type GetStreamTabResponse = {

	/**Title of this page */
	title: string;

	type: TabType;

	style: TabStyle;

	showNames?: boolean;
};